+++ cmd/kubectl-mongodb/migrate-to-mck/replica_set_generator.go
@@ -6,11 +6,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	k8svalidation "k8s.io/apimachinery/pkg/util/validation"
 
 	mdbv1 "github.com/mongodb/mongodb-kubernetes/api/v1/mdb"
+	mdbmulti "github.com/mongodb/mongodb-kubernetes/api/v1/mdbmulti"
+	userv1 "github.com/mongodb/mongodb-kubernetes/api/v1/user"
 	"github.com/mongodb/mongodb-kubernetes/controllers/om"
-	"github.com/mongodb/mongodb-kubernetes/pkg/util"
 )
 
 func generateReplicaSet(ac *om.AutomationConfig, opts GenerateOptions) (client.Object, string, error) {
@@ -23,17 +23,19 @@ func generateReplicaSet(ac *om.AutomationConfig, opts GenerateOptions) (client.O
 	rsName := rs.Name()
 	externalMembers, version, fcv := om.ExtractMemberInfo(rs.Members(), ac.Deployment.ProcessMap())
 
-	resourceName := opts.ResourceNameOverride
+	resourceName := opts.ReplicaSetNameOverride
 	if resourceName == "" {
-		resourceName = util.NormalizeName(rsName)
-		if resourceName == "" {
-			return nil, "", fmt.Errorf("replica set name %q cannot be normalized to a valid Kubernetes resource name. Use --resource-name-override to provide one", rsName)
+		if userv1.NormalizeName(rsName) != rsName {
+			return nil, "", fmt.Errorf("replica set name %q is not a valid Kubernetes resource name. Use --replicaset-name-override to provide a valid name (spec.replicaSetNameOverride will be set automatically)", rsName)
 		}
-	}
-	if errs := k8svalidation.IsDNS1123Subdomain(resourceName); len(errs) > 0 {
-		return nil, "", fmt.Errorf("resource name %q is not a valid Kubernetes resource name: %s", resourceName, errs[0])
+		resourceName = rsName
+	} else if userv1.NormalizeName(resourceName) != resourceName {
+		return nil, "", fmt.Errorf("--replicaset-name-override value %q is not a valid Kubernetes resource name", resourceName)
 	}
 
+	if len(opts.MultiClusterNames) > 0 {
+		return generateReplicaSetMultiCluster(ac, opts, rsName, resourceName, version, fcv, externalMembers)
+	}
 	return generateReplicaSetSingleCluster(ac, opts, rsName, resourceName, version, fcv, externalMembers)
 }
 
@@ -49,11 +51,23 @@ func generateReplicaSetSingleCluster(ac *om.AutomationConfig, opts GenerateOptio
 	}, resourceName, nil
 }
 
+func generateReplicaSetMultiCluster(ac *om.AutomationConfig, opts GenerateOptions, rsName, resourceName, version, fcv string, externalMembers []mdbv1.ExternalMember) (client.Object, string, error) {
+	spec, err := buildReplicaSetMultiClusterSpec(ac, opts, version, fcv, externalMembers, rsName, resourceName)
+	if err != nil {
+		return nil, "", fmt.Errorf("failed to build multi-cluster spec: %w", err)
+	}
+	return &mdbmulti.MongoDBMultiCluster{
+		TypeMeta:   metav1.TypeMeta{APIVersion: "mongodb.com/v1", Kind: "MongoDBMultiCluster"},
+		ObjectMeta: buildCRObjectMeta(resourceName, opts.Namespace),
+		Spec:       spec,
+	}, resourceName, nil
+}
 
 // buildReplicaSetDbCommonSpec constructs the DbCommonSpec for a replica set deployment,
 // including security, Prometheus, TLS, and connection settings.
 func buildReplicaSetDbCommonSpec(ac *om.AutomationConfig, opts GenerateOptions, version, fcv, rsName, resourceName string, externalMembers []mdbv1.ExternalMember) (mdbv1.DbCommonSpec, error) {
-	security, err := buildSecurity(ac, opts.CertsSecretPrefix, resourceName)
+	rs := ac.Deployment.GetReplicaSets()[0]
+	security, err := buildSecurity(ac.Auth, ac.Deployment.ProcessMap(), rs.Members(), ac.Ldap, ac.OIDCProviderConfigs, opts.CertsSecretPrefix)
 	if err != nil {
 		return mdbv1.DbCommonSpec{}, fmt.Errorf("failed to build security config: %w", err)
 	}
@@ -112,6 +126,23 @@ func buildReplicaSetSpec(ac *om.AutomationConfig, opts GenerateOptions, version,
 	}
 	return mdbv1.MongoDbSpec{
 		DbCommonSpec: common,
-		Members:      0,
+		Members:      len(externalMembers),
+		MemberConfig: buildMemberConfig(ac.Deployment.GetReplicaSets()[0].Members()),
+	}, nil
+}
+
+// buildReplicaSetMultiClusterSpec assembles a MongoDBMultiSpec, distributing members across target clusters.
+func buildReplicaSetMultiClusterSpec(ac *om.AutomationConfig, opts GenerateOptions, version, fcv string, externalMembers []mdbv1.ExternalMember, rsName, resourceName string) (mdbmulti.MongoDBMultiSpec, error) {
+	common, err := buildReplicaSetDbCommonSpec(ac, opts, version, fcv, rsName, resourceName, externalMembers)
+	if err != nil {
+		return mdbmulti.MongoDBMultiSpec{}, err
+	}
+	clusterSpecList, err := distributeMembers(externalMembers, ac.Deployment.GetReplicaSets()[0].Members(), opts.MultiClusterNames)
+	if err != nil {
+		return mdbmulti.MongoDBMultiSpec{}, err
+	}
+	return mdbmulti.MongoDBMultiSpec{
+		DbCommonSpec:    common,
+		ClusterSpecList: clusterSpecList,
 	}, nil
 }
