+++ cmd/kubectl-mongodb/migrate-to-mck/user_generator.go
@@ -6,7 +6,9 @@ import (
 
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	k8svalidation "k8s.io/apimachinery/pkg/util/validation"
 
 	userv1 "github.com/mongodb/mongodb-kubernetes/api/v1/user"
 	"github.com/mongodb/mongodb-kubernetes/controllers/om"
@@ -14,8 +16,9 @@ import (
 )
 
 // GenerateUserCRs creates MongoDBUser CRs for each user in the automation config, skipping the agent user.
-// Each SCRAM user references a pre-created Secret from ExistingUserSecrets.
-// External users (X.509 and LDAP) never produce a Secret.
+// When ExistingUserSecrets is set, each SCRAM user references a pre-created Secret, otherwise new Secrets
+// are generated from UserPasswords. External users (X509 and LDAP) never produce a Secret.
+// Output is ordered with each MongoDBUser followed by its Secret when one is generated.
 func GenerateUserCRs(ac *om.AutomationConfig, mongodbResourceName, namespace string, opts GenerateOptions) ([]client.Object, error) {
 	if ac.Auth == nil || len(ac.Auth.Users) == 0 {
 		return nil, nil
@@ -40,7 +43,7 @@ func GenerateUserCRs(ac *om.AutomationConfig, mongodbResourceName, namespace str
 			return nil, fmt.Errorf("username %q cannot be normalized to a valid Kubernetes name: no alphanumeric characters", user.Username)
 		}
 		if prev, exists := crNameToUsername[crName]; exists {
-			return nil, fmt.Errorf("users %q and %q normalize to the same Kubernetes name %q. Rename one before migration", prev, user.Username, crName)
+			return nil, fmt.Errorf("users %q and %q normalize to the same Kubernetes name %q; rename one before migration", prev, user.Username, crName)
 		}
 		crNameToUsername[crName] = user.Username
 
@@ -58,13 +61,28 @@ func GenerateUserCRs(ac *om.AutomationConfig, mongodbResourceName, namespace str
 			Roles: roles,
 		}
 
+		var passwordSecret *corev1.Secret
 		if user.Database != externalDatabase {
-			sName, ok := opts.ExistingUserSecrets[userKey(user.Username, user.Database)]
-			if !ok {
-				fmt.Fprintf(os.Stderr, "[WARNING] skipping user %q (db: %s): no Secret name provided\n", user.Username, user.Database)
-				continue
+			if opts.ExistingUserSecrets != nil {
+				sName, ok := opts.ExistingUserSecrets[userKey(user.Username, user.Database)]
+				if !ok {
+					fmt.Fprintf(os.Stderr, "[WARNING] skipping user %q (db: %s): not found in --users-secrets-file\n", user.Username, user.Database)
+					continue
+				}
+				spec.PasswordSecretKeyRef = userv1.SecretKeyRef{Name: sName, Key: passwordSecretDataKey}
+			} else {
+				passwordSecretName := crName + "-password"
+				if errs := k8svalidation.IsDNS1123Subdomain(passwordSecretName); len(errs) > 0 {
+					return nil, fmt.Errorf("generated password Secret name %q is not a valid Kubernetes name; rename user %q before migration: %s", passwordSecretName, user.Username, errs[0])
+				}
+				password, ok := opts.UserPasswords[userKey(user.Username, user.Database)]
+				if !ok {
+					fmt.Fprintf(os.Stderr, "[WARNING] skipping user %q (db: %s): no password provided\n", user.Username, user.Database)
+					continue
+				}
+				spec.PasswordSecretKeyRef = userv1.SecretKeyRef{Name: passwordSecretName, Key: passwordSecretDataKey}
+				passwordSecret = GeneratePasswordSecret(passwordSecretName, namespace, password)
 			}
-			spec.PasswordSecretKeyRef = userv1.SecretKeyRef{Name: sName, Key: passwordSecretDataKey}
 		}
 
 		results = append(results, &userv1.MongoDBUser{
@@ -72,6 +90,9 @@ func GenerateUserCRs(ac *om.AutomationConfig, mongodbResourceName, namespace str
 			ObjectMeta: metav1.ObjectMeta{Name: crName, Namespace: namespace},
 			Spec:       spec,
 		})
+		if passwordSecret != nil {
+			results = append(results, passwordSecret)
+		}
 	}
 
 	return results, nil
